server-py/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-socketio==5.11.3
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==3.2.2
